+++ lib/src/utils/isolate_utils.dart
@@ -1,41 +1,43 @@
 import 'dart:isolate';
 import 'dart:math';
+import 'dart:typed_data';
 
 import 'package:camera/camera.dart';
 
-import 'flutter_zxing.dart';
+import '../../generated_bindings.dart';
+import '../logic/zxing.dart';
+import 'image_converter.dart';
 
 // Inspired from https://github.com/am15h/object_detection_flutter
 
 /// Bundles data to pass between Isolate
 class IsolateData {
-  CameraImage cameraImage;
-  int format;
-  double cropPercent;
-
-  SendPort? responsePort;
-
   IsolateData(
     this.cameraImage,
     this.format,
     this.cropPercent,
   );
+  CameraImage cameraImage;
+  int format;
+  double cropPercent;
+
+  SendPort? responsePort;
 }
 
 /// Manages separate Isolate instance for inference
 class IsolateUtils {
-  static const String kDebugName = "ZxingIsolate";
+  static const String kDebugName = 'ZxingIsolate';
 
   // ignore: unused_field
   Isolate? _isolate;
-  final _receivePort = ReceivePort();
+  final ReceivePort _receivePort = ReceivePort();
   SendPort? _sendPort;
 
   SendPort? get sendPort => _sendPort;
 
-  Future<void> start() async {
+  Future<void> startReadingBarcode() async {
     _isolate = await Isolate.spawn<SendPort>(
-      entryPoint,
+      readBarcodeEntryPoint,
       _receivePort.sendPort,
       debugName: kDebugName,
     );
@@ -43,30 +45,36 @@ class IsolateUtils {
     _sendPort = await _receivePort.first;
   }
 
-  void stop() {
+  void stopReadingBarcode() {
     _isolate?.kill(priority: Isolate.immediate);
     _isolate = null;
     _sendPort = null;
   }
 
-  static void entryPoint(SendPort sendPort) async {
-    final port = ReceivePort();
+  static Future<void> readBarcodeEntryPoint(SendPort sendPort) async {
+    final ReceivePort port = ReceivePort();
     sendPort.send(port.sendPort);
 
     await for (final IsolateData? isolateData in port) {
       if (isolateData != null) {
         try {
-          final image = isolateData.cameraImage;
-          final cropPercent = isolateData.cropPercent;
-          final bytes = await convertImage(image);
-          final cropSize =
+          final CameraImage image = isolateData.cameraImage;
+          final double cropPercent = isolateData.cropPercent;
+          final Uint8List bytes = await convertImage(image);
+          final int cropSize =
               (min(image.width, image.height) * cropPercent).round();
 
-          final result = FlutterZxing.readBarcode(bytes, isolateData.format,
-              image.width, image.height, cropSize, cropSize);
+          final CodeResult result = readBarcode(
+            bytes,
+            width: image.width,
+            height: image.height,
+            format: isolateData.format,
+            cropWidth: cropSize,
+            cropHeight: cropSize,
+          );
 
           isolateData.responsePort?.send(result);
-        } on Exception catch (e) {
+        } catch (e) {
           isolateData.responsePort?.send(e);
         }
       }