+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## 0.4.0
+Breaking changes
+* removed `FlutterZxing` class, call all methods directly
+* added read multiple barcodes methods
+
+## 0.3.2
+
+* fixed enabling/disabling of the logger
+
+## 0.3.1
+
+* fixed Chinese support
+
+## 0.3.0
+
+* added processCameraImage function
+* added pinch to zoom sopport
+* added flash sopport
+* added custom scanner overlay support
+
+## 0.2.0
+
+* added 'readImagePath' function
+* added 'readImagePathString' function
+* added 'readImageUrl' function
+
+## 0.1.3
+
+* minor improvements
+
 ## 0.1.2
 
 * minor fixes for analyzer options